organized code - deleted old unused variables and moved functions to shapes.cpp and tools.cpp files
