+++ Paint Panther/include/apiHandler.h
@@ -1,10 +1,8 @@
-// ApiHandler.h
 #ifndef API_HANDLER_H
 #define API_HANDLER_H
 
 #include <string>
 
-// Declare the API-related functions
 std::string getResponse();
 
-#endif // API_HANDLER_H
+#endif
